+++ components/module/example/demo-overlay.tsx
@@ -1,5 +1,7 @@
 "use client";
 import CButton from "@/components/shared/custome/c-button";
+import CDialog from "@/components/shared/custome/c-dialog";
+import CDrawer from "@/components/shared/custome/c-drawer";
 import store from "@/store";
 
 export default function DemoOverlay() {
@@ -48,6 +50,29 @@ export default function DemoOverlay() {
         onClick={() => handleConfirmDisableInteraction(false)}
         variant="secondary"
       />
+      <CDrawer
+        trigger={
+          <CButton
+            title="using children and trigger inside comp"
+            variant="secondary"
+          />
+        }
+      >
+        Lorem ipsum dolor sit amet, qui minim labore adipisicing minim sint
+        cillum sint consectetur cupidatat.
+      </CDrawer>
+      <CDialog
+        title={<p>helo</p>}
+        trigger={
+          <CButton
+            title="using children and trigger inside comp"
+            variant="secondary"
+          />
+        }
+      >
+        Lorem ipsum dolor sit amet, qui minim labore adipisicing minim sint
+        cillum sint consectetur cupidatat.
+      </CDialog>
     </div>
   );
 }