+++ components/shared/overlay-wrapper/index.tsx
@@ -1,19 +1,9 @@
 import * as React from "react";
-
-import { Drawer, DrawerContent, DrawerHeader } from "@/components/ui/drawer";
-
-import {
-  Dialog,
-  DialogContent,
-  DialogHeader,
-  DialogTitle,
-} from "@/components/ui/dialog";
 import { useBreakpoint } from "@/lib/hooks";
 import store from "@/store";
-import { cn } from "@/lib/utils";
-import CloseOverlay from "../close-overlay";
-import { VisuallyHidden } from "@radix-ui/react-visually-hidden";
 import OvConfirmation from "./ov-confirmation";
+import CDrawer from "../custome/c-drawer";
+import CDialog from "../custome/c-dialog";
 
 export default function OverlayWrapper() {
   const { isMobile } = useBreakpoint();
@@ -27,10 +17,6 @@ export default function OverlayWrapper() {
   const titleAlign = overlay?.titleAlign || "start";
   const disableOutsideInteraction = overlay?.disableOutsideInteraction || false;
 
-  const handleDisableOutsideInteraction = (e: KeyboardEvent | any) => {
-    disableOutsideInteraction && e.preventDefault();
-  };
-
   const Content = () => {
     switch (id) {
       case "CONFRIMATION":
@@ -38,64 +24,17 @@ export default function OverlayWrapper() {
     }
   };
 
-  return isMobile ? (
-    <Drawer
-      open={open}
-      onOpenChange={() => closeOverlay()}
-      dismissible={!disableOutsideInteraction}
-    >
-      <DrawerContent>
-        {title ? (
-          <DrawerHeader>
-            <DialogTitle
-              className={classNameTitle}
-              style={{ textAlign: titleAlign }}
-            >
-              {title}
-            </DialogTitle>
-            {isClose && <CloseOverlay />}
-          </DrawerHeader>
-        ) : (
-          <VisuallyHidden asChild>
-            <DialogTitle></DialogTitle>
-          </VisuallyHidden>
-        )}
-        <div
-          className={cn(
-            "max-h-[80vh] w-full max-w-xl mx-auto",
-            isPadding ? "p-[16px]" : "p-0",
-          )}
-        >
-          {Content()}
-        </div>
-      </DrawerContent>
-    </Drawer>
-  ) : (
-    <Dialog open={open} onOpenChange={() => closeOverlay()}>
-      <DialogContent
-        className={cn(isPadding ? "p-[16px]" : "p-0")}
-        onInteractOutside={(e) => handleDisableOutsideInteraction(e)}
-        onEscapeKeyDown={(e) => handleDisableOutsideInteraction(e)}
-        aria-describedby={title ? title : undefined}
-      >
-        {title ? (
-          <DialogHeader>
-            <DialogTitle
-              className={classNameTitle}
-              style={{ textAlign: titleAlign }}
-            >
-              {title}
-            </DialogTitle>
-            {isClose && <CloseOverlay />}
-          </DialogHeader>
-        ) : (
-          <VisuallyHidden asChild>
-            <DialogTitle></DialogTitle>
-          </VisuallyHidden>
-        )}
-        {!title && isClose && <CloseOverlay />}
-        {Content()}
-      </DialogContent>
-    </Dialog>
-  );
+  const sharedProps = {
+    open,
+    title,
+    isClose,
+    onClose: closeOverlay,
+    classNameTitle,
+    titleAlign,
+    isPadding,
+    disableOutsideInteraction,
+    children: <Content />,
+  };
+
+  return isMobile ? <CDrawer {...sharedProps} /> : <CDialog {...sharedProps} />;
 }