+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 all notable changes to this project will be documented in this file.
 
+# [0.1.0] - 2025-06-28
+
+- (Feat) - cradio and cchecbox
+- (Feat) - cradio-form and ccheckbox-form
+- (Refactor) - move all input field to folder form/
+
 # [0.1.0] - 2025-06-26
 
 - (Feat) - put dialog and drawer (become overlay wrapper)